requirements.txt

chardet
openai==0.28
streamlit==1.37.0
requests==2.32.3
PyGithub
